refactor: ToyShop package 변경
